Clean up scripts and add shield counter state logic
