+++ Orbit Pong/Assets/Scripts/PaddleBarrier.cs
