+++ Orbit Pong/Assets/Sprites.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 2cda990e2423bbf4892e6590ba056729
+guid: 64d8c60262912b441881d64410c39a6f
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
