+++ Orbit Pong/Assets/Sprites/ShieldPaddles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a527c8cb1753f4b528e243c044a3c413
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
