+++ Orbit Pong/Assets/Sprites/ShieldPaddles/Fire-Shield-First-Charge.png.meta
