+++ Orbit Pong/Assets/Sprites/ShieldPaddles/Fire-Shield-Full-Charge.png.meta
