+++ Orbit Pong/Assets/Sprites/ShieldPaddles/Fire-Shield-Half-Charge.png
