+++ Orbit Pong/Assets/Sprites/ShieldPaddles/Fire-Shield-No-Charge.png
