+++ Orbit Pong/UserSettings/EditorUserSettings.asset
@@ -6,16 +6,28 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
-      flags: 0
-    RecentlyUsedScenePath-1:
       value: 224247031146464b1f01182d043a4f15562149651f2b233e283a5227e7f32076f7e93ffdfe
       flags: 0
+    RecentlyUsedScenePath-1:
+      value: 224247031146466b1f01182d043a4f1556214a651f2b233e283a5227e7f32076f7e93ffdfe
+      flags: 0
     RecentlyUsedScenePath-2:
-      value: 224247031146466b1f01182d043a4f15562149651f2b233e283a5207ebf4383dd1e424ece2347129370bfb25
+      value: 224247031146466b1f01182d043a4f15562149651f2b233e283a5227e7f32076f7e93ffdfe
       flags: 0
     RecentlyUsedScenePath-3:
+      value: 224247031146466b1f01182d043a4f15562149651f2b233e283a5207ebf4383dd1e424ece2347129370bfb25
+      flags: 0
+    RecentlyUsedScenePath-4:
       value: 224247031146466b1f01182d043a4f15562149651f2b233e283a521cf0e23d2cc0e624fbee3f2d722c0ce6281d
+      flags: 0
+    RecentlyUsedScenePath-5:
+      value: 22424703114646680e0b0227036c4b150503563f22213229
+      flags: 0
+    RecentlyUsedScenePath-6:
+      value: 22424703114646680e0b0227036c7002141e0c082d3a3439283b5326ece92021
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c6c111b07142f1f2b233e2867083debf42d
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
